+++ src/Basket/Controller/Basket.php
@@ -18,6 +18,7 @@ use OxidEsales\GraphQL\Storefront\Payment\DataType\BasketPayment;
 use TheCodingMachine\GraphQLite\Annotations\Logged;
 use TheCodingMachine\GraphQLite\Annotations\Mutation;
 use TheCodingMachine\GraphQLite\Annotations\Query;
+use TheCodingMachine\GraphQLite\Annotations\Right;
 use TheCodingMachine\GraphQLite\Types\ID;
 
 final class Basket
@@ -46,7 +47,7 @@ final class Basket
 
     /**
      * @Mutation()
-     * @Logged()
+     * @Right("ADD_PRODUCT_TO_BASKET")
      */
     public function basketAddProduct(string $basketId, string $productId, float $amount): BasketDataType
     {
@@ -55,7 +56,7 @@ final class Basket
 
     /**
      * @Mutation()
-     * @Logged()
+     * @Right("REMOVE_BASKET_PRODUCT")
      */
     public function basketRemoveProduct(string $basketId, string $productId, int $amount): BasketDataType
     {
@@ -64,7 +65,7 @@ final class Basket
 
     /**
      * @Mutation()
-     * @Logged()
+     * @Right("CREATE_BASKET")
      */
     public function basketCreate(BasketDataType $basket): BasketDataType
     {
@@ -116,7 +117,7 @@ final class Basket
 
     /**
      * @Mutation()
-     * @Logged()
+     * @Right("ADD_VOUCHER")
      */
     public function basketAddVoucher(string $basketId, string $voucherNumber): BasketDataType
     {
@@ -125,7 +126,7 @@ final class Basket
 
     /**
      * @Mutation()
-     * @Logged()
+     * @Right("REMOVE_VOUCHER")
      */
     public function basketRemoveVoucher(string $basketId, string $voucherId): BasketDataType
     {
@@ -185,7 +186,7 @@ final class Basket
 
     /**
      * @Mutation()
-     * @Logged()
+     * @Right("PLACE_ORDER")
      */
     public function placeOrder(ID $basketId, ?bool $confirmTermsAndConditions = null): OrderDataType
     {