+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Remove ``final`` statement from shop extending classes [PR-3](https://github.com/OXID-eSales/graphql-storefront-module/pull/3).
 
+### Changed
+- Use Rights annotation instead of Logged so that an anonymus token user could create/modify basket and place an order.
+
 ## [0.1.0] - 2020-12-16
 
 - Initial release